+++ pom.xml
@@ -8,14 +8,14 @@
     <version>1.0.0</version>
 
     <parent>
-       <groupId>io.lippia</groupId>
-       <artifactId>root</artifactId>
-       <version>3.2.3.8</version>
+        <groupId>io.lippia</groupId>
+        <artifactId>root</artifactId>
+        <version>3.2.3.8</version>
     </parent>
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <crowdar.cucumber.filter>@Smoke</crowdar.cucumber.filter>
+        <crowdar.cucumber.filter>@Demo</crowdar.cucumber.filter>
         <cucumber.runner>testngParallel.xml</cucumber.runner>
         <crowdar.driverHub></crowdar.driverHub>
         <crowdar.cucumber.option>src/test/resources/features --glue com/crowdar --glue com/crowdar/bdd/cukes  --tags ${crowdar.cucumber.filter} --tags 'not @Ignore' --plugin com.crowdar.examples.report.Reporter: --plugin pretty</crowdar.cucumber.option>
@@ -27,27 +27,39 @@
     <profiles>
         <profile>
             <id>Local</id>
-            <properties>
-                <crowdar.driverHub>http://127.0.0.1:4723/wd/hub</crowdar.driverHub>
-                <cucumber.runner>testng.xml</cucumber.runner>
-                <crowdar.avd>Pixel3_Testing</crowdar.avd>
-                <crowdar.appLocation>C:\\\\Users\\\\germa\\\\OneDrive\\\\Documentos\\\\Crowdar\\\\8-9-TrabajoPractico-LippiaMobile\\\\lippia-mobile-sample-project\\\\sample_apk\\\\lippia.apk</crowdar.appLocation>
-                <crowdar.deviceName>Pixel3</crowdar.deviceName>
-            </properties>
-        </profile>
-        <profile>
-            <id>Clockify</id>
             <activation>
                 <activeByDefault>true</activeByDefault>
             </activation>
             <properties>
                 <crowdar.driverHub>http://127.0.0.1:4723/wd/hub</crowdar.driverHub>
                 <cucumber.runner>testng.xml</cucumber.runner>
+                <crowdar.appLocation>sample_apk/Clockify.apk</crowdar.appLocation>
+            </properties>
+        </profile>
+
+        <profile>
+            <id>Emulado</id>
+            <activation>
+                <activeByDefault>true</activeByDefault>
+            </activation>
+            <properties>
                 <crowdar.avd>Pixel3_Testing</crowdar.avd>
-                <crowdar.appLocation>C:\\\\Users\\\\germa\\\\OneDrive\\\\Documentos\\\\Crowdar\\\\8-9-TrabajoPractico-LippiaMobile\\\\lippia-mobile-sample-project\\\\sample_apk\\\\Clockify.apk</crowdar.appLocation>
+                <crowdar.projectType.driverCapabilities.jsonFile>src/main/resources/capabilities/emuladorCapabilities.json</crowdar.projectType.driverCapabilities.jsonFile>
                 <crowdar.deviceName>Pixel3</crowdar.deviceName>
             </properties>
         </profile>
+
+        <profile>
+            <id>Fisico</id>
+            <activation>
+                <activeByDefault>false</activeByDefault>
+            </activation>
+            <properties>
+                <crowdar.projectType.driverCapabilities.jsonFile>src/main/resources/capabilities/fisicoCapabilities.json</crowdar.projectType.driverCapabilities.jsonFile>
+                <crowdar.deviceName>GalaxyA32</crowdar.deviceName>
+            </properties>
+        </profile>
+
         <profile>
             <id>StackLocal</id>
             <properties>
@@ -72,7 +84,7 @@
                 <crowdar.projectType>MOBILE_IOS</crowdar.projectType>
             </properties>
         </profile>
-    	<profile>
+        <profile>
             <id>BrowserStack</id>
             <properties>
                 <cucumber.runner>testng.xml</cucumber.runner>
@@ -87,9 +99,9 @@
 
     <dependencies>
         <dependency>
-                <groupId>io.lippia</groupId>
-                <artifactId>core</artifactId>
-                <version>${project.parent.version}</version>
+            <groupId>io.lippia</groupId>
+            <artifactId>core</artifactId>
+            <version>${project.parent.version}</version>
         </dependency>
         <dependency>
             <groupId>io.lippia.report</groupId>
@@ -110,8 +122,8 @@
         <plugins>
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
-				<artifactId>maven-surefire-plugin</artifactId>
-				<version>3.0.0-M4</version>
+                <artifactId>maven-surefire-plugin</artifactId>
+                <version>3.0.0-M4</version>
                 <configuration>
                     <useSystemClassLoader>false</useSystemClassLoader>
                     <suiteXmlFiles>
@@ -123,11 +135,11 @@
     </build>
 
     <repositories>
-		<repository>
-			<id>crowdarRepo</id>
-			<name>crowdar-repository</name>
-			<url>https://nexus-v3-repositories.automation.crowdaronline.com/repository/maven-public/</url>
-		</repository>
-	</repositories>
+        <repository>
+            <id>crowdarRepo</id>
+            <name>crowdar-repository</name>
+            <url>https://nexus-v3-repositories.automation.crowdaronline.com/repository/maven-public/</url>
+        </repository>
+    </repositories>
 
 </project>
