--- src/main/java/com/crowdar/examples/constants/ClockifyLoginConstants.java
@@ -1,14 +1,0 @@
-package com.crowdar.examples.constants;
-
-public class ClockifyLoginConstants {
-    public static final String APP_VISIBLE = "xpath://android.widget.ImageView[@content-desc='Clockify logo']";
-
-    public static final String USER = "id:emailAddressInput";
-
-    public static final String CONTRASENA = "id:passwordInput";
-
-    public static final String LOGIN = "id:loginButton";
-
-    public static final String HOME_APP = "xpath://android.widget.TextView[@content-desc='Time entries']";
-
-}
