+++ src/main/java/com/crowdar/examples/constants/CrearUnEscenarioConstants.java
@@ -3,7 +3,7 @@ package com.crowdar.examples.constants;
 public class CrearUnEscenarioConstants {
     public static final String AGREGAR_ESCENARIO = "xpath://android.widget.ImageButton[@content-desc='Add time entry']";
     public static final String AGREGAR_TIEMPO = "id:constraintLayout";
-    public static final String DIA = "xpath://android.view.View[@content-desc='08 October 2023']";
+    public static final String DIA = "xpath://android.view.View[@content-desc='%s']";
     public static final String INICIO = "xpath://android.widget.LinearLayout[@content-desc='START']";
     public static final String FIN = "xpath://android.widget.LinearLayout[@content-desc='END']";
     public static final String SELECIONAR_TIEMPO = "xpath://android.widget.NumberPicker[1]/android.widget.EditText";
