--- src/main/java/com/crowdar/examples/constants/HomeConstants.java
@@ -1,8 +1,0 @@
-package com.crowdar.examples.constants;
-
-public class HomeConstants {
-
-    public static final String SIGN_OUT_BUTTON_LOCATOR = "ACCESSIBILITY_ID:signOutButtonHome";
-    public static final String CHANGE_LANGUAGE_BUTTON_LOCATOR = "ACCESSIBILITY_ID:changeLanguageButtonHome";
-    public static final String VIEW_NOT_DISPLAYED_MESSAGE = "Home page is not displayed.";
-}
