+++ src/main/java/com/crowdar/examples/constants/LoginConstants.java
@@ -1,11 +1,14 @@
 package com.crowdar.examples.constants;
 
 public class LoginConstants {
+    public static final String APP_VISIBLE = "xpath://android.widget.ImageView[@content-desc='Clockify logo']";
 
-    public static final String EMAIL_INPUT_LOCATOR = "ACCESSIBILITY_ID:emailAddressInputLogin";
-    public static final String PASSWORD_INPUT_LOCATOR = "ACCESSIBILITY_ID:passwordInputLogin";
-    public static final String SIGN_IN_BUTTON_LOCATOR = "ACCESSIBILITY_ID:signInButtonLogin";
-    public static final String SIGN_UP_BUTTON_LOCATOR = "ACCESSIBILITY_ID:signUpButtonLogin";
+    public static final String USER = "id:emailAddressInput";
 
-    public static final String VIEW_NOT_DISPLAYED_MESSAGE = "Login page is not displayed.";
+    public static final String CONTRASENA = "id:passwordInput";
+
+    public static final String LOGIN = "id:loginButton";
+
+    public static final String HOME_APP = "xpath://android.widget.TextView[@content-desc='Time entries']";
+
 }
