--- src/main/java/com/crowdar/examples/constants/SignUpConstants.java
@@ -1,9 +1,0 @@
-package com.crowdar.examples.constants;
-
-public class SignUpConstants {
-
-    public static final String USERNAME_INPUT_LOCATOR = "ACCESSIBILITY_ID:usernameInputSignUp";
-    public static final String EMAIL_INPUT_LOCATOR = "ACCESSIBILITY_ID:emailAddressInputSignUp";
-    public static final String PASSWORD_INPUT_LOCATOR = "ACCESSIBILITY_ID:passwordInputSignUp";
-    public static final String SIGN_UP_BUTTON_LOCATOR = "ACCESSIBILITY_ID:signUpButtonSignUp";
-}
