+++ src/main/java/com/crowdar/examples/services/LoginService.java
@@ -5,19 +5,18 @@ import com.crowdar.examples.constants.LoginConstants;
 import org.testng.Assert;
 
 public class LoginService {
-
-    public static void hacerLogin(String email, String password){
-        MobileActionManager.setInput(LoginConstants.EMAIL_INPUT_LOCATOR, email);
-        MobileActionManager.setInput(LoginConstants.PASSWORD_INPUT_LOCATOR, password);
-        MobileActionManager.click(LoginConstants.SIGN_IN_BUTTON_LOCATOR);
+    public static void verificacionApp() {
+        MobileActionManager.waitVisibility(LoginConstants.APP_VISIBLE);
     }
 
-    public static void irRegistro(){
-        MobileActionManager.click(LoginConstants.SIGN_UP_BUTTON_LOCATOR);
+    public static void datosLogin(String user, String contrasena) {
+        MobileActionManager.setInput(LoginConstants.USER,user);
+        MobileActionManager.setInput(LoginConstants.CONTRASENA,contrasena);
+        MobileActionManager.click(LoginConstants.LOGIN);
     }
 
-    public static void isViewLoaded(){
-        MobileActionManager.waitVisibility(LoginConstants.SIGN_UP_BUTTON_LOCATOR);
-        Assert.assertTrue(MobileActionManager.isVisible(LoginConstants.EMAIL_INPUT_LOCATOR), LoginConstants.VIEW_NOT_DISPLAYED_MESSAGE);
+    public static void ingresoApp() {
+        MobileActionManager.waitVisibility(LoginConstants.HOME_APP);
+        Assert.assertTrue(MobileActionManager.isVisible(LoginConstants.HOME_APP));
     }
 }
